// backend/src/constants/errors/error.enum.ts
export enum ErrorCode {
  SUCCESS = 0,
  
  // Auth & General
  UNAUTHORIZED = 401001,
  FORBIDDEN = 403001,
  NOT_FOUND = 404001,
  INTERNAL_SERVER_ERROR = 500001,
  VALIDATION_ERROR = 400001,

  // User
  USER_NOT_FOUND = 404101,
  USER_ALREADY_EXISTS = 400101,
  USER_BLOCKED = 403101,
  USER_BORROW_LIMIT_EXCEEDED = 400102,

  // Book
  BOOK_NOT_FOUND = 404201,
  BOOK_NOT_AVAILABLE = 400201,
  BOOK_ALREADY_EXISTS = 400202,

  // Borrow
  BORROW_RECORD_NOT_FOUND = 404301,
  INVALID_BORROW_OPERATION = 400301,
}